requirements.txt
extra_streamlit_components==0.1.60
numpy==1.23.5
pandas==2.2.1
openpyxl==3.1.0
plotly==5.9.0
python_dateutil==2.8.2
streamlit==1.29.0
